feat: fix lab-5

+++ src/cart/cart.service.ts
@@ -2,36 +2,41 @@ import { Injectable } from "@nestjs/common";
 import { CartDto } from "./dto/cart.dto";
 import { CreateCartRequest } from "./dto/createCartRequest";
 import { Cart } from "@prisma/client";
-import { RuntimeException } from "@nestjs/core/errors/exceptions";
 import { PrismaService } from "../prisma.service";
+import {
+  CartAlreadyExistsException,
+  CartNotFoundException,
+  InvalidCreateCartRequestException
+} from "./utils/exceptions";
 
 @Injectable()
 export class CartService {
 
   constructor(
-    private prisma: PrismaService,
+    private prisma: PrismaService
   ) {
   }
 
   async getCart(cartId: string): Promise<CartDto> {
-    const cart = await this.prisma.cart.findFirst({
+    const cart = await this.prisma.cart.findUnique({
       where: {
         id: Number(cartId)
       }
     });
+    if (cart == null) {
+      throw new CartNotFoundException("no cart with such id");
+    }
     const items = await this.prisma.cartProductItem.findMany({
       where: {
         cartId: cart.id
       }
-    })
+    });
 
     return this.toCartDto(cart, items.map((value) => value.shopProductItemId.toString()));
   }
 
   async createCart(createCartRequest: CreateCartRequest): Promise<CartDto> {
-    if (!this.validateRequest()) {
-      throw new RuntimeException("invalid request input");
-    }
+    await this.validateRequest(createCartRequest);
 
     const time = new Date().toISOString();
     const cart = await this.prisma.cart.create({
@@ -40,10 +45,15 @@ export class CartService {
         userId: createCartRequest.userId
       }
     });
+
     return this.toCartDto(cart, []);
   }
 
   async deleteCart(cartId: string): Promise<boolean> {
+    if (await this.prisma.cart.findUnique({ where: { id: Number(cartId) } }) == null) {
+      throw new CartNotFoundException("no cart with such id");
+    }
+
     const deletedCart = await this.prisma.cart.delete({
       where: {
         id: Number(cartId)
@@ -59,6 +69,10 @@ export class CartService {
         userId: Number(userId)
       }
     });
+
+    if (cart === null) {
+      throw new CartNotFoundException(`no cart for such user = ${userId}`);
+    }
 
     const cartProductItemIds: string[] = (await this.prisma.cartProductItem.findMany({
       where: {
@@ -76,11 +90,17 @@ export class CartService {
       cart.id,
       cart.userId,
       cart.createdAt.toString(),
-      cartProductItemIds,
+      cartProductItemIds
     );
   }
 
-  private validateRequest(): boolean {
-    return true;
+  private async validateRequest(createCartRequest: CreateCartRequest) {
+    if (await this.prisma.user.findUnique({ where: { id: createCartRequest.userId } }) === null) {
+      throw new InvalidCreateCartRequestException("no such user registered");
+    }
+
+    if (await this.prisma.cart.findUnique({ where: { id: createCartRequest.userId } }) !== null) {
+      throw new CartAlreadyExistsException("cart for this user already exists");
+    }
   }
 }
