+++ src/order/order.service.ts
@@ -1,13 +1,14 @@
 import { Injectable } from "@nestjs/common";
 import { OrderDto } from "./dto/order.dto";
 import { CreateOrderRequest } from "./dto/createOrderRequest";
-import { UserDto } from "../user/dto/user.dto";
 import { PrismaService } from "../prisma.service";
-import { UserService } from "../user/user.service";
 import { Order, OrderLine } from "@prisma/client";
-import { RuntimeException } from "@nestjs/core/errors/exceptions";
-import { ProductsService } from "../products/products.service";
 import { ProductDto } from "../products/dto/product.dto";
+import {
+  InvalidCreateOrderRequestException,
+  OrderAlreadyExistsException,
+  OrderNotFoundException
+} from "./utils/exceptions";
 
 @Injectable()
 export class OrderService {
@@ -21,6 +22,11 @@ export class OrderService {
         id: Number(orderId)
       }
     });
+
+    if (order == null) {
+      throw new OrderNotFoundException(`no order with such orderId= ${orderId}`);
+    }
+
     const orderLines = await this.prisma.orderLine.findMany({
       where: {
         orderId: Number(order.id)
@@ -31,23 +37,22 @@ export class OrderService {
   }
 
   async createOrder(createOrderRequest: CreateOrderRequest): Promise<OrderDto> {
-    if (!this.validateRequest(createOrderRequest)) {
-      throw new RuntimeException("invalid request input");
-    }
+    await this.validateRequest(createOrderRequest);
+    console.log("ORDERLINES", createOrderRequest.orderLines);
 
-    const createdOrderLines: OrderLine[] = [];
-    for (const orderLine of createOrderRequest.orderLines) {
-      const product = await this.getProduct(orderLine.orderLineProductId.toString());
-      createdOrderLines.push(await this.prisma.orderLine.create({
-        data: {
-          orderId: orderLine.orderId,
-          tax: orderLine.tax,
-          totalAmount: orderLine.totalAmount,
-          discount: orderLine.discount,
-          productName: product.title
-        }
-      }));
-    }
+    // const createdOrderLines: OrderLine[] = [];
+    // createOrderRequest.orderLines.map(async orderLine => {
+    //   const product = await this.getProduct(orderLine.orderLineProductId.toString());
+    //   createdOrderLines.push(await this.prisma.orderLine.create({
+    //     data: {
+    //       orderId: orderLine.orderId,
+    //       tax: orderLine.tax,
+    //       totalAmount: orderLine.totalAmount,
+    //       discount: orderLine.discount,
+    //       productName: product.title
+    //     }
+    //   }));
+    // })
 
     const time = new Date().toISOString();
     const order = await this.prisma.order.create({
@@ -55,13 +60,17 @@ export class OrderService {
         createdAt: time,
         totalPrice: createOrderRequest.totalPrice,
         cartId: createOrderRequest.cartId,
-        userId: createOrderRequest.userId,
+        userId: createOrderRequest.userId
       }
     });
-    return this.toOrderDto(order, createdOrderLines);
+    return this.toOrderDto(order, []);
   }
 
   async deleteOrder(orderId: string): Promise<boolean> {
+    if (await this.prisma.order.findUnique({ where: { id: Number(orderId) } }) == null) {
+      throw new OrderNotFoundException(`no order with such orderId= ${orderId}`);
+    }
+
     const deletedUser = await this.prisma.order.delete({
       where: {
         id: Number(orderId)
@@ -95,8 +104,8 @@ export class OrderService {
     return new OrderDto(
       order.id,
       order.createdAt.toString(),
-      order.userId,
       order.totalPrice,
+      order.userId,
       orderLines.map(function(value) {
         return value.productName;
       }),
@@ -129,7 +138,17 @@ export class OrderService {
     );
   }
 
-  private validateRequest(createOrderRequest: CreateOrderRequest): boolean {
-    return true;
+  private async validateRequest(createOrderRequest: CreateOrderRequest) {
+    const userId = createOrderRequest.userId;
+    const cartId = createOrderRequest.cartId;
+    if (await this.prisma.user.findUnique({ where: { id: userId } }) === null) {
+      throw new InvalidCreateOrderRequestException("no such user registered");
+    }
+    if (await this.prisma.cart.findUnique({ where: { id: cartId } }) === null) {
+      throw new InvalidCreateOrderRequestException("no such cartId exists");
+    }
+    if (await this.prisma.order.findUnique({ where: { cartId: cartId } }) !== null) {
+      throw new OrderAlreadyExistsException("order for this user and cart already exists");
+    }
   }
 }
